chore: More refactors
