+++ package/src/Camera.tsx
@@ -7,7 +7,7 @@ import type { CameraProps, OnShutterEvent } from './types/CameraProps'
 import { CameraModule } from './NativeCameraModule'
 import type { PhotoFile, TakePhotoOptions } from './types/PhotoFile'
 import type { Point } from './types/Point'
-import type { RecordVideoOptions, VideoFile } from './VideoFile'
+import type { RecordVideoOptions, VideoFile } from './types/VideoFile'
 import { VisionCameraProxy } from './FrameProcessorPlugins'
 import { CameraDevices } from './CameraDevices'
 import type { EmitterSubscription, NativeSyntheticEvent, NativeMethods } from 'react-native'
