+++ package/src/index.ts
@@ -1,15 +1,16 @@
 export * from './Camera'
-export * from './types/CameraDevice'
 export * from './CameraError'
+export * from './FrameProcessorPlugins'
+
+export * from './types/CameraDevice'
 export * from './types/CameraProps'
 export * from './types/Frame'
-export * from './FrameProcessorPlugins'
 export * from './types/Orientation'
 export * from './types/PhotoFile'
 export * from './types/Snapshot'
 export * from './types/PixelFormat'
 export * from './types/Point'
-export * from './VideoFile'
+export * from './types/VideoFile'
 export * from './types/CodeScanner'
 
 export * from './devices/getCameraFormat'
@@ -20,7 +21,7 @@ export * from './hooks/useCameraDevice'
 export * from './hooks/useCameraDevices'
 export * from './hooks/useCameraFormat'
 export * from './hooks/useCameraPermission'
-export * from './hooks/useFrameProcessor'
 export * from './hooks/useCodeScanner'
+export * from './hooks/useFrameProcessor'
 
 export * from './skia/useSkiaFrameProcessor'
