+++ package/src/types/PhotoFile.ts
@@ -1,5 +1,5 @@
 import type { Orientation } from './Orientation'
-import type { TemporaryFile } from '../TemporaryFile'
+import type { TemporaryFile } from './TemporaryFile'
 
 export interface TakePhotoOptions {
   /**
